added scale/music theory fundamentals plus tests

+++ src/clarinoid/application/Metronome.hpp
@@ -7,13 +7,13 @@ struct Metronome
     CCThrottlerT<500> mMetronomeTimer; // the 500 is arbitrary
     int GetBeatInt() const {
         // gSynthGraphControl.mMetronomeTimer.GetBeatFloat(60000.0f / gAppSettings.mBPM)
-        return mMetronomeTimer.GetBeatInt(60000.0f / gAppSettings.mBPM);
+        return mMetronomeTimer.GetBeatInt((int)(60000.0f / gAppSettings.mBPM));
     }
     float GetBeatFloat() const {
-        return mMetronomeTimer.GetBeatFloat(60000.0f / gAppSettings.mBPM);
+        return mMetronomeTimer.GetBeatFloat((int)(60000.0f / gAppSettings.mBPM));
     }
     float GetBeatFrac() const {
-        return mMetronomeTimer.GetBeatFrac(60000.0f / gAppSettings.mBPM);
+        return mMetronomeTimer.GetBeatFrac((int)(60000.0f / gAppSettings.mBPM));
     }
 };
